Rename tests accordingly
