+++ StepLang.Tooling.Formatting.Tests/Analyzers/LineEndingAnalyzerTest.cs
@@ -1,14 +1,14 @@
 using StepLang.Tooling.Formatting.Analyzers;
 
-namespace StepLang.Tooling.Formatting.Tests.Fixers;
+namespace StepLang.Tooling.Formatting.Tests.Analyzers;
 
-public class LineEndingFixerTest
+public class LineEndingAnalyzerTest
 {
     [Theory]
     [InlineData("\r\n", "\n")]
     [InlineData("\n", "\n")]
     [InlineData("\r\n\r\n", "\n\n")]
-    public async Task TestLineEndingFixerFixesLineEndings(string input, string output)
+    public async Task TestLineEndingAnalyzer(string input, string output)
     {
         var fixer = new LineEndingAnalyzer();
